First commit
This is a simple but useful data structure for string reference counting.
It was part of a bachelor thesis, where the goal was to create a software, which is monitoring the users keyboard behavior.
After monitoring it builds up a profile for the current user and detects the differences.
Anyway, this data structure stored the most used words by the user.